+++ src/g_canvas.c
@@ -331,20 +331,20 @@ t_canvas *canvas_new(void *dummy, t_symbol *sel, int argc, t_atom *argv)
 
     if (argc == 5)  /* toplevel: x, y, w, h, font */
     {
-        xloc = (int)atom_getintarg(0, argc, argv);
-        yloc = (int)atom_getintarg(1, argc, argv);
-        width = (int)atom_getintarg(2, argc, argv);
-        height = (int)atom_getintarg(3, argc, argv);
-        font = (int)atom_getintarg(4, argc, argv);
+        xloc = atom_getintarg(0, argc, argv);
+        yloc = atom_getintarg(1, argc, argv);
+        width = atom_getintarg(2, argc, argv);
+        height = atom_getintarg(3, argc, argv);
+        font = atom_getintarg(4, argc, argv);
     }
     else if (argc == 6)  /* subwindow: x, y, w, h, name, vis */
     {
-        xloc = (int)atom_getintarg(0, argc, argv);
-        yloc = (int)atom_getintarg(1, argc, argv);
-        width = (int)atom_getintarg(2, argc, argv);
-        height = (int)atom_getintarg(3, argc, argv);
+        xloc = atom_getintarg(0, argc, argv);
+        yloc = atom_getintarg(1, argc, argv);
+        width = atom_getintarg(2, argc, argv);
+        height = atom_getintarg(3, argc, argv);
         s = atom_getsymbolarg(4, argc, argv);
-        vis = (int)atom_getintarg(5, argc, argv);
+        vis = atom_getintarg(5, argc, argv);
     }
         /* (otherwise assume we're being created from the menu.) */
     if (THISGUI->i_newdirectory &&
@@ -409,15 +409,15 @@ static void canvas_coords(t_glist *x, t_symbol *s, int argc, t_atom *argv)
     x->gl_y1 = atom_getfloatarg(1, argc, argv);
     x->gl_x2 = atom_getfloatarg(2, argc, argv);
     x->gl_y2 = atom_getfloatarg(3, argc, argv);
-    x->gl_pixwidth = (int)atom_getintarg(4, argc, argv);
-    x->gl_pixheight = (int)atom_getintarg(5, argc, argv);
+    x->gl_pixwidth = atom_getintarg(4, argc, argv);
+    x->gl_pixheight = atom_getintarg(5, argc, argv);
     if (argc <= 7)
-        canvas_setgraph(x, (int)atom_getintarg(6, argc, argv), 1);
+        canvas_setgraph(x, atom_getintarg(6, argc, argv), 1);
     else
     {
-        x->gl_xmargin = (int)atom_getintarg(7, argc, argv);
-        x->gl_ymargin = (int)atom_getintarg(8, argc, argv);
-        canvas_setgraph(x, (int)atom_getintarg(6, argc, argv), 0);
+        x->gl_xmargin = atom_getintarg(7, argc, argv);
+        x->gl_ymargin = atom_getintarg(8, argc, argv);
+        canvas_setgraph(x, atom_getintarg(6, argc, argv), 0);
     }
 }
 
@@ -1207,7 +1207,7 @@ void glob_dsp(void *dummy, t_symbol *s, int argc, t_atom *argv)
     int newstate;
     if (argc)
     {
-        newstate = (int)atom_getintarg(0, argc, argv);
+        newstate = atom_getintarg(0, argc, argv);
         if (newstate && !THISGUI->i_dspstate)
         {
             sys_set_audio_state(1);
