+++ src/g_hdial.c
@@ -291,9 +291,9 @@ static void hradio_properties(t_gobj *z, t_glist *owner)
 static void hradio_dialog(t_hradio *x, t_symbol *s, int argc, t_atom *argv)
 {
     t_symbol *srl[3];
-    int a = (int)atom_getintarg(0, argc, argv);
-    int chg = (int)atom_getintarg(4, argc, argv);
-    int num = (int)atom_getintarg(6, argc, argv);
+    int a = atom_getintarg(0, argc, argv);
+    int chg = atom_getintarg(4, argc, argv);
+    int num = atom_getintarg(6, argc, argv);
     int sr_flags;
 
     if(chg != 0) chg = 1;
@@ -512,7 +512,7 @@ static void hradio_number(t_hradio *x, t_floatarg num)
 
 static void hradio_size(t_hradio *x, t_symbol *s, int ac, t_atom *av)
 {
-    x->x_gui.x_w = iemgui_clip_size((int)atom_getintarg(0, ac, av));
+    x->x_gui.x_w = iemgui_clip_size(atom_getintarg(0, ac, av));
     x->x_gui.x_h = x->x_gui.x_w;
     iemgui_size((void *)x, &x->x_gui);
 }
@@ -577,15 +577,15 @@ static void *hradio_donew(t_symbol *s, int argc, t_atom *argv, int old)
        &&IS_A_FLOAT(argv,7)&&IS_A_FLOAT(argv,8)
        &&IS_A_FLOAT(argv,9)&&IS_A_FLOAT(argv,10)&&IS_A_FLOAT(argv,14))
     {
-        a = (int)atom_getintarg(0, argc, argv);
-        chg = (int)atom_getintarg(1, argc, argv);
-        iem_inttosymargs(&x->x_gui.x_isa, (int)atom_getintarg(2, argc, argv));
-        num = (int)atom_getintarg(3, argc, argv);
+        a = atom_getintarg(0, argc, argv);
+        chg = atom_getintarg(1, argc, argv);
+        iem_inttosymargs(&x->x_gui.x_isa, atom_getintarg(2, argc, argv));
+        num = atom_getintarg(3, argc, argv);
         iemgui_new_getnames(&x->x_gui, 4, argv);
-        ldx = (int)atom_getintarg(7, argc, argv);
-        ldy = (int)atom_getintarg(8, argc, argv);
-        iem_inttofstyle(&x->x_gui.x_fsf, (int)atom_getintarg(9, argc, argv));
-        fs = (int)atom_getintarg(10, argc, argv);
+        ldx = atom_getintarg(7, argc, argv);
+        ldy = atom_getintarg(8, argc, argv);
+        iem_inttofstyle(&x->x_gui.x_fsf, atom_getintarg(9, argc, argv));
+        fs = atom_getintarg(10, argc, argv);
         iemgui_all_loadcolors(&x->x_gui, argv+11, argv+12, argv+13);
         fval = atom_getfloatarg(14, argc, argv);
     }
