+++ src/g_numbox.c
@@ -485,12 +485,12 @@ static void my_numbox_dialog(t_my_numbox *x, t_symbol *s, int argc,
     t_atom *argv)
 {
     t_symbol *srl[3];
-    int w = (int)atom_getintarg(0, argc, argv);
-    int h = (int)atom_getintarg(1, argc, argv);
+    int w = atom_getintarg(0, argc, argv);
+    int h = atom_getintarg(1, argc, argv);
     double min = (double)atom_getfloatarg(2, argc, argv);
     double max = (double)atom_getfloatarg(3, argc, argv);
-    int lilo = (int)atom_getintarg(4, argc, argv);
-    int log_height = (int)atom_getintarg(6, argc, argv);
+    int lilo = atom_getintarg(4, argc, argv);
+    int log_height = atom_getintarg(6, argc, argv);
     int sr_flags;
 
     if(lilo != 0) lilo = 1;
@@ -604,13 +604,13 @@ static void my_numbox_size(t_my_numbox *x, t_symbol *s, int ac, t_atom *av)
 {
     int h, w;
 
-    w = (int)atom_getintarg(0, ac, av);
+    w = atom_getintarg(0, ac, av);
     if(w < 1)
         w = 1;
     x->x_gui.x_w = w;
     if(ac > 1)
     {
-        h = (int)atom_getintarg(1, ac, av);
+        h = atom_getintarg(1, ac, av);
         if(h < 8)
             h = 8;
         x->x_gui.x_h = h;
@@ -653,12 +653,12 @@ static void my_numbox_label_pos(t_my_numbox *x, t_symbol *s, int ac, t_atom *av)
 static void my_numbox_label_font(t_my_numbox *x,
     t_symbol *s, int ac, t_atom *av)
 {
-    int f = (int)atom_getintarg(1, ac, av);
+    int f = atom_getintarg(1, ac, av);
 
     if(f < 4)
         f = 4;
     x->x_gui.x_fontsize = f;
-    f = (int)atom_getintarg(0, ac, av);
+    f = atom_getintarg(0, ac, av);
     if((f < 0) || (f > 2))
         f = 0;
     x->x_gui.x_fsf.x_font_style = f;
@@ -775,24 +775,24 @@ static void *my_numbox_new(t_symbol *s, int argc, t_atom *argv)
        &&IS_A_FLOAT(argv,9)&&IS_A_FLOAT(argv,10)
        &&IS_A_FLOAT(argv,11)&&IS_A_FLOAT(argv,12)&&IS_A_FLOAT(argv,16))
     {
-        w = (int)atom_getintarg(0, argc, argv);
-        h = (int)atom_getintarg(1, argc, argv);
+        w = atom_getintarg(0, argc, argv);
+        h = atom_getintarg(1, argc, argv);
         min = (double)atom_getfloatarg(2, argc, argv);
         max = (double)atom_getfloatarg(3, argc, argv);
-        lilo = (int)atom_getintarg(4, argc, argv);
-        iem_inttosymargs(&x->x_gui.x_isa, (int)atom_getintarg(5, argc, argv));
+        lilo = atom_getintarg(4, argc, argv);
+        iem_inttosymargs(&x->x_gui.x_isa, atom_getintarg(5, argc, argv));
         iemgui_new_getnames(&x->x_gui, 6, argv);
-        ldx = (int)atom_getintarg(9, argc, argv);
-        ldy = (int)atom_getintarg(10, argc, argv);
-        iem_inttofstyle(&x->x_gui.x_fsf, (int)atom_getintarg(11, argc, argv));
-        fs = (int)atom_getintarg(12, argc, argv);
+        ldx = atom_getintarg(9, argc, argv);
+        ldy = atom_getintarg(10, argc, argv);
+        iem_inttofstyle(&x->x_gui.x_fsf, atom_getintarg(11, argc, argv));
+        fs = atom_getintarg(12, argc, argv);
         iemgui_all_loadcolors(&x->x_gui, argv+13, argv+14, argv+15);
         v = atom_getfloatarg(16, argc, argv);
     }
     else iemgui_new_getnames(&x->x_gui, 6, 0);
     if((argc == 18)&&IS_A_FLOAT(argv,17))
     {
-        log_height = (int)atom_getintarg(17, argc, argv);
+        log_height = atom_getintarg(17, argc, argv);
     }
     x->x_gui.x_draw = (t_iemfunptr)my_numbox_draw;
     x->x_gui.x_fsf.x_snd_able = 1;
